# api_docs.yaml
openapi: 3.0.0
info:
  title: LaunchPad Placement Portal API
  version: 1.0.0
  description: |
    REST API for the LaunchPad campus placement portal.

    **Authentication:** Most endpoints require an `Authentication-Token` header.
    Log in via `/auth/login` to get your token, then click "Authorize" above and paste it.
servers:
  - url: http://localhost:5001/api
paths:
...
  /student/profile:
    get:
      summary: Get student profile
      tags: [Student]
      security:
        - TokenAuth: []
      responses:
        "200": { description: Student profile data }
    put:
      summary: Update student profile
      tags: [Student]
      security:
        - TokenAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentProfile"
            example:
              name: Alok Kumar
              branch: Computer Science
              cgpa: 9.2
              phone: "+91 9876543210"
              skills: Python, Flask, Vue.js
              experience: "2 internships at startups"
      responses:
        "200": { description: Profile updated }
  /student/resume:
    get:
      summary: Download resume
      tags: [Student]
      security:
        - TokenAuth: []
      responses:
        "200":
          description: PDF file download
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "404": { description: No resume uploaded }
    post:
      summary: Upload resume (PDF only)
      tags: [Student]
      security:
        - TokenAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [resume]
              properties:
                resume:
                  type: string
                  format: binary
                  description: PDF file
      responses:
        "200": { description: Resume uploaded }
        "400": { description: No file or invalid format }
  /student/drives:
    get:
      summary: Browse open placement drives
      description: Returns approved, open drives the student is eligible for.
      tags: [Student]
      security:
        - TokenAuth: []
      parameters:
        - name: search
          in: query
          schema: { type: string }
          description: Search by title, role, or company name
        - name: branch
          in: query
          schema: { type: string }
          description: Filter by branch eligibility
        - name: min_package
          in: query
          schema: { type: number }
          description: Minimum package in LPA
      responses:
        "200": { description: Eligible drives list }
  /student/apply/{drive_id}:
    post:
      summary: Apply to a placement drive
      description: ATS score is computed automatically from resume and drive description.
      tags: [Student]
      security:
        - TokenAuth: []
      parameters:
        - { name: drive_id, in: path, required: true, schema: { type: integer } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                cover_letter: { type: string }
            example:
              cover_letter: Excited about this opportunity!
      responses:
        "201": { description: Application submitted }
        "400": { description: Ineligible or deadline passed }
        "409": { description: Already applied }
  /student/applications:
    get:
      summary: Track applications
      tags: [Student]
      security:
        - TokenAuth: []
      responses:
        "200": { description: Student's applications list }
  /student/applications/{app_id}/withdraw:
    put:
      summary: Withdraw application
      description: Only applications with status "applied" can be withdrawn.
      tags: [Student]
      security:
        - TokenAuth: []
      parameters:
        - { name: app_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Application withdrawn }
        "400": { description: Cannot withdraw (wrong status) }
  /student/applications/export:
    post:
      summary: Export applications as CSV via email
      description: Triggers a Celery task. CSV is sent to the student's email.
      tags: [Student]
      security:
        - TokenAuth: []
      responses:
        "200": { description: Export task queued }
  /student/placements:
    get:
      summary: View secured placements
      tags: [Student]
      security:
        - TokenAuth: []
      responses:
        "200": { description: Student's placements list }
components:
  securitySchemes:
    TokenAuth:
      type: apiKey
      in: header
      name: Authentication-Token
      description: Paste the token returned by POST /auth/login
  schemas:
...
    UpdateStudentProfile:
      type: object
      properties:
        name: { type: string }
        branch: { type: string }
        cgpa: { type: number }
        phone: { type: string }
        skills: { type: string }
        experience: { type: string }
...